requirements.txt
django
djangorestframework
requests
PyJWT
pytest 
pytest-django 
requests-mock
pytest-cov
django-cors-headers